Added minor changes with solution

+++ ECommerce/ECommerce.Api.Orders/Models/OrderModel.cs
@@ -8,7 +8,7 @@ namespace ECommerce.Api.Orders.Models
         public int Id { get; set; }
         public int CustomerId { get; set; }
         public DateTime OrderDate { get; set; }
-        public int Total { get; set; }
+        public decimal Total { get; set; }
         public OrderItem[] Items { get; set; }
     }
 }
